requirements.txt
fastapi
uvicorn
httpx
ccxt
pandas
numpy
python-dotenv
pydantic
mysql-connector-python
firebase-admin
metaapi-cloud-sdk
aiosqlite
playwright
crawl4ai
beautifulsoup4
rich
pydantic-settings
redis
python-socketio
chromadb
python-telegram-bot
